+++ packages/twenty-sdk/src/application/application-config.ts
@@ -1,14 +1,1 @@
-import { type SyncableEntityOptions } from '@/application/syncable-entity-options.type';
-
-type ApplicationVariable = SyncableEntityOptions & {
-  value?: string;
-  description?: string;
-  isSecret?: boolean;
-};
-
-export type ApplicationConfig = SyncableEntityOptions & {
-  displayName?: string;
-  description?: string;
-  icon?: string;
-  applicationVariables?: Record<string, ApplicationVariable>;
-};
+export { type Application as ApplicationConfig } from 'twenty-shared/application';
