+++ packages/twenty-sdk/src/cli/utils/load-manifest.ts
@@ -40,7 +40,6 @@ import {
   type ApplicationManifest,
   type ServerlessFunctionManifest,
   type ObjectManifest,
-  type FieldManifest,
 } from 'twenty-shared/application';
 import { findPathFile } from '@/cli/utils/find-path-file';
 import { getTsProgramAndDiagnostics } from '@/cli/utils/get-ts-program-and-diagnostics';
@@ -185,7 +184,7 @@ const collectObjects = (program: Program) => {
               }
 
               fields.push({
-                ...(fieldCfg as FieldManifest),
+                ...(fieldCfg as any),
                 ...(name ? { name } : {}),
               });
             }
