+++ packages/twenty-shared/src/application/applicationVariablesType.ts
@@ -1,5 +1,6 @@
-type ApplicationVariable = {
-  universalIdentifier: string;
+import { type SyncableEntityOptions } from '@/application/syncableEntityOptionsType';
+
+type ApplicationVariable = SyncableEntityOptions & {
   value?: string;
   description?: string;
   isSecret?: boolean;
