+++ packages/twenty-shared/src/application/fieldManifestType.ts
@@ -1,14 +1,23 @@
-import { type FieldMetadataType } from '@/types';
+import {
+  type FieldMetadataType,
+  type FieldMetadataSettings,
+  type FieldMetadataOptions,
+  type FieldMetadataDefaultValue,
+} from '@/types';
+import { type SyncableEntityOptions } from '@/application/syncableEntityOptionsType';
 
-export type FieldManifest = {
-  universalIdentifier: string;
-  type: FieldMetadataType;
+export type FieldManifest<
+  T extends FieldMetadataType = Exclude<
+    FieldMetadataType,
+    FieldMetadataType.RELATION
+  >,
+> = SyncableEntityOptions & {
+  type: T;
   label: string;
   description?: string;
   icon?: string;
-  defaultValue?: any;
-  options?: any;
-  settings?: any;
+  defaultValue?: FieldMetadataDefaultValue<T>;
+  options?: FieldMetadataOptions<T>;
+  settings?: FieldMetadataSettings<T>;
   isNullable?: boolean;
-  isFieldUiReadOnly?: boolean;
 };
