+++ packages/twenty-shared/src/application/objectManifestType.ts
@@ -1,7 +1,7 @@
 import { type FieldManifest } from '@/application';
+import { type SyncableEntityOptions } from '@/application/syncableEntityOptionsType';
 
-export type ObjectManifest = {
-  universalIdentifier: string;
+export type ObjectManifest = SyncableEntityOptions & {
   nameSingular: string;
   namePlural: string;
   labelSingular: string;
