+++ packages/twenty-shared/src/application/syncableEntityOptionsType.ts
@@ -1,0 +1,1 @@
+export type SyncableEntityOptions = { universalIdentifier: string };
